// src/sections/activities/types.ts
export interface ActivitesData {
  Region: string;
  "First Name": string;
  "Corporate ID": string;
  "E-mail": string;
  Country: string;
  "Operating Group": string;
  Franchise: string;
  Sector: string;
  Manager: string;
  Date: string;
  Collision: string;
  Classification: string;
  Status: string;
}
export interface ActivitesDataPIFS {
  Region: string;
  "First Name": string;
  "Corporate ID": string;
  "E-mail": string;
  Country: string;
  "Operating Group": string;
  Franchise: string;
  Sector: string;
  Manager: string;
  Date: string;
  Collision: string;
  Classification: string;
  Status: string;
}